+++ core/internal/randomid/randomtid_test.go
@@ -1,15 +1,15 @@
-package utils_test
+package randomid_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/wandb/wandb/core/pkg/utils"
+	"github.com/wandb/wandb/core/internal/randomid"
 )
 
 func TestShortID(t *testing.T) {
 	t.Run("shortid", func(t *testing.T) {
-		item := utils.ShortID(8)
+		item := randomid.GenerateUniqueID(8)
 		assert.Equal(t, len(item), 8)
 	})
 }
