+++ core/pkg/utils/utils.go
@@ -1,11 +1,5 @@
 package utils
 
-import (
-	"math/rand"
-)
-
-const alphanumericChars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-
 func NilIfZero[T comparable](x T) *T {
 	var zero T
 	if x == zero {
@@ -21,14 +15,4 @@ func ZeroIfNil[T comparable](x *T) T {
 		return zero
 	}
 	return *x
-}
-
-func GenerateAlphanumericSequence(length int) string {
-	var result string
-	for i := 0; i < length; i++ {
-		index := rand.Intn(len(alphanumericChars))
-		result += string(alphanumericChars[index])
-	}
-
-	return result
 }
